# documentation/openapi.yaml
openapi: 3.1.0
info:
  title: Toolvio Backend API - RBAC & Dynamic Schema System
  version: 1.0.0
  description: |
    Comprehensive API documentation for Toolvio Backend with Role-Based Access Control (RBAC),
    Dynamic Schema Management, Audit Trail, and Multi-tenant Support.
    
    ## 🔐 Authentication
    Most endpoints require JWT authentication. Use the `/api/auth/login` endpoint to get a token,
    then include it in the Authorization header as `Bearer <token>`.
    
    ## 👥 User Roles
    - **Admin**: Full access to all features
    - **Office**: Can manage schemas and data, read audit logs
    - **Technician**: Can read schemas, manage data, read audit logs
    - **Customer**: Can only read data (no schema or audit access)
    
    ## 🏢 Multi-tenant
    All data is isolated by tenant. Users can only access data within their assigned tenant.
    
    Use the "Try it out" button to execute requests against your local server.
servers:
  - url: http://localhost:{port}
    description: Local server
    variables:
      port:
        default: "3000"
paths:
...
  /api/auth/tenant:
    get:
      tags: [Authentication]
      summary: Get tenant information
      description: Retrieve current user's tenant information
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Tenant information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/TenantInfo'
                  message:
                    type: string
        '401':
          description: Authentication required
        '404':
          description: Tenant not found
...
  /api/tenants:
    post:
      tags: [Tenants]
      summary: Create a new tenant
      description: Create a new tenant with the specified configuration
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantCreate'
      responses:
        '201':
          description: Tenant created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/TenantResponse'
                  message:
                    type: string
        '400':
          description: Invalid request data or tenant already exists
        '401':
          description: Authentication required
        '403':
          description: Insufficient permissions
    get:
      tags: [Tenants]
      summary: Get all tenants
      description: Retrieve all tenants with pagination, filtering, and sorting
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of items per page
        - in: query
          name: search
          schema:
            type: string
          description: Search term for name, display name, or tenant ID
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive]
          description: Filter by tenant status
        - in: query
          name: plan
          schema:
            type: string
            enum: [trial, basic, professional, enterprise]
          description: Filter by subscription plan
        - in: query
          name: sortBy
          schema:
            type: string
            default: createdAt
          description: Field to sort by
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
      responses:
        '200':
          description: Tenants retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      tenants:
                        type: array
                        items:
                          $ref: '#/components/schemas/TenantResponse'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: number
                          limit:
                            type: number
                          total:
                            type: number
                          pages:
                            type: number
                  message:
                    type: string
        '401':
          description: Authentication required
        '403':
          description: Insufficient permissions
  /api/tenants/summary:
    get:
      tags: [Tenants]
      summary: Get tenants summary
      description: Get a simplified list of active tenants for dropdowns and selection
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Tenants summary retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        tenantId:
                          type: string
                        name:
                          type: string
                        displayName:
                          type: string
                        subscriptionPlan:
                          type: string
                  message:
                    type: string
        '401':
          description: Authentication required
  /api/tenants/{tenantId}:
    get:
      tags: [Tenants]
      summary: Get tenant by ID
      description: Retrieve a specific tenant by its ID
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: string
          description: Tenant identifier
      responses:
        '200':
          description: Tenant retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/TenantResponse'
                  message:
                    type: string
        '401':
          description: Authentication required
        '404':
          description: Tenant not found
    put:
      tags: [Tenants]
      summary: Update tenant
      description: Update an existing tenant's information
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: string
          description: Tenant identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantUpdate'
      responses:
        '200':
          description: Tenant updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/TenantResponse'
                  message:
                    type: string
        '400':
          description: Invalid request data
        '401':
          description: Authentication required
        '403':
          description: Insufficient permissions
        '404':
          description: Tenant not found
    delete:
      tags: [Tenants]
      summary: Delete tenant (soft delete)
      description: Deactivate a tenant (soft delete - sets isActive to false)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: string
          description: Tenant identifier
      responses:
        '200':
          description: Tenant deactivated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/TenantResponse'
                  message:
                    type: string
        '401':
          description: Authentication required
        '403':
          description: Insufficient permissions
        '404':
          description: Tenant not found
  /api/tenants/{tenantId}/status:
    patch:
      tags: [Tenants]
      summary: Toggle tenant status
      description: Activate or deactivate a tenant
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: string
          description: Tenant identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - isActive
              properties:
                isActive:
                  type: boolean
                  description: Whether to activate or deactivate the tenant
                  example: true
      responses:
        '200':
          description: Tenant status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/TenantResponse'
                  message:
                    type: string
        '400':
          description: Invalid request data
        '401':
          description: Authentication required
        '403':
          description: Insufficient permissions
        '404':
          description: Tenant not found
  /api/tenants/{tenantId}/stats:
    get:
      tags: [Tenants]
      summary: Get tenant statistics
      description: Retrieve usage statistics and limits for a specific tenant
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: string
          description: Tenant identifier
      responses:
        '200':
          description: Tenant statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/TenantStats'
                  message:
                    type: string
        '401':
          description: Authentication required
        '404':
          description: Tenant not found
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT


      description: JWT token from /api/auth/login endpoint
...
  schemas:
...
    TenantInfo:
      type: object
      properties:
        _id:
          type: string
          description: Tenant ID
        tenantId:
          type: string
          description: Tenant identifier
        name:
          type: string
          description: Tenant name
        displayName:
          type: string
          description: Display name
        description:
          type: string
          description: Tenant description
        subscriptionPlan:
          type: string
          enum: [trial, basic, professional, enterprise]
          description: Subscription plan
        subscriptionStatus:
          type: string
          description: Current subscription status
        settings:
          type: object
          description: Tenant settings
        usage:
          type: object
          description: Usage statistics

    # Sync-related schemas
...